Add Invoke-HardenedUNCPathCheck
The vulnerability patched by MS15-011 allows a man in the middle to
inject group policies when the DC is polled for updates, because Windows
used to not check the file shares identity by default. Starting with
Windows 10, the file shares identity is indeed checked by default. They
call this "hardened UNC path". However, the hardening had to be
explicitely configured for SYSVOL before Windows 10. Also,
Administrators can disable the UNC path hardening manually. If they do,
privilege escalation to SYSTEM is possible for anyone in a man in
the middle position.

This commit adds the Cmdlet Invoke-HardenedUNCPathCheck which reports
whether the system is vulnerable.

I rate this medium because a legit DC has to be available and
exploitation can be difficult. I have managed to so in the past, but I'm
not aware of *convenient* public exploit.